existing check with s3

+++ after/upload-s3.rb
@@ -14,6 +14,7 @@ S3_BUCKET = Aws::S3::Resource.new.bucket(ENV['S3_BUCKET'])
 Dir::glob("after/thumbs/**/*.jpg").each do |path|
   key = path.sub 'after', 'items'
   obj = S3_BUCKET.object(key)
+  next if obj.exists?
   result = obj.upload_file(path)
   ap "#{key} => #{result}"
   File.delete(path)
